planner: refactor the Task and move new package
