+++ pkg/planner/core/operator/physicalop/task/task.go
@@ -1,0 +1,7 @@
+package task
+
+import "github.com/pingcap/tidb/pkg/planner/core/base"
+
+var (
+	_ base.Task = &RootTask{}
+)
