+++ pkg/planner/core/operator/physicalop/task/util.go
@@ -1,0 +1,167 @@
+package task
+
+import (
+	"bytes"
+	"fmt"
+	"unsafe"
+
+	"github.com/pingcap/tidb/pkg/expression"
+	"github.com/pingcap/tidb/pkg/parser/mysql"
+	"github.com/pingcap/tidb/pkg/planner/util/utilfuncp"
+	"github.com/pingcap/tidb/pkg/types"
+	"github.com/pingcap/tidb/pkg/util/chunk"
+	"github.com/pingcap/tidb/pkg/util/ranger"
+	rangerctx "github.com/pingcap/tidb/pkg/util/ranger/context"
+	"github.com/pingcap/tidb/pkg/util/size"
+)
+
+// IndexJoinInfo is generated by index join's inner ds, which will build their own index choice based
+// the indexJoinProp pushed down by index join. While index join still need some feedback by this kind
+// choice info to make index join runtime compatible, like IdxColLens will help truncate the index key
+// to construct suitable lookup contents. KeyOff2IdxOff will help ds to quickly locate the index column
+// from lookup contents. Ranges will be used to rebuild the underlying index range if there is any parameter
+// affecting the index join's inner range. CompareFilters will be used to quickly evaluate the last-col's
+// non-eq range.
+// This kind of IndexJoinInfo will be wrapped as a part of CopTask or RootTask, which will be passed upward
+// to targeted indexJoin to complete the physicalIndexJoin's detail: ref:
+type IndexJoinInfo struct {
+	// The following fields are used to keep index join aware of inner plan's index/pk choice.
+	IdxColLens     []int
+	KeyOff2IdxOff  []int
+	Ranges         ranger.MutableRanges
+	CompareFilters *ColWithCmpFuncManager
+}
+
+// ColWithCmpFuncManager is used in index join to handle the column with compare functions(>=, >, <, <=).
+// It stores the compare functions and build ranges in execution phase.
+type ColWithCmpFuncManager struct {
+	TargetCol         *expression.Column
+	ColLength         int
+	OpType            []string
+	opArg             []expression.Expression
+	TmpConstant       []*expression.Constant
+	AffectedColSchema *expression.Schema  `plan-cache-clone:"shallow"`
+	compareFuncs      []chunk.CompareFunc `plan-cache-clone:"shallow"`
+}
+
+func (cwc *ColWithCmpFuncManager) CloneForPlanCache() *ColWithCmpFuncManager {
+	if cwc == nil {
+		return nil
+	}
+	cloned := new(ColWithCmpFuncManager)
+	if cwc.TargetCol != nil {
+		if cwc.TargetCol.SafeToShareAcrossSession() {
+			cloned.TargetCol = cwc.TargetCol
+		} else {
+			cloned.TargetCol = cwc.TargetCol.Clone().(*expression.Column)
+		}
+	}
+	cloned.ColLength = cwc.ColLength
+	cloned.OpType = make([]string, len(cwc.OpType))
+	copy(cloned.OpType, cwc.OpType)
+	cloned.opArg = utilfuncp.CloneExpressionsForPlanCache(cwc.opArg, nil)
+	cloned.TmpConstant = utilfuncp.CloneConstantsForPlanCache(cwc.TmpConstant, nil)
+	cloned.AffectedColSchema = cwc.AffectedColSchema
+	cloned.compareFuncs = cwc.compareFuncs
+	return cloned
+}
+
+// AppendNewExpr appends a new expression to ColWithCmpFuncManager.
+func (cwc *ColWithCmpFuncManager) AppendNewExpr(opName string, arg expression.Expression, affectedCols []*expression.Column) {
+	cwc.OpType = append(cwc.OpType, opName)
+	cwc.opArg = append(cwc.opArg, arg)
+	cwc.TmpConstant = append(cwc.TmpConstant, &expression.Constant{RetType: cwc.TargetCol.RetType})
+	for _, col := range affectedCols {
+		if cwc.AffectedColSchema.Contains(col) {
+			continue
+		}
+		cwc.compareFuncs = append(cwc.compareFuncs, chunk.GetCompareFunc(col.RetType))
+		cwc.AffectedColSchema.Append(col)
+	}
+}
+
+// CompareRow compares the rows for deduplicate.
+func (cwc *ColWithCmpFuncManager) CompareRow(lhs, rhs chunk.Row) int {
+	for i, col := range cwc.AffectedColSchema.Columns {
+		ret := cwc.compareFuncs[i](lhs, col.Index, rhs, col.Index)
+		if ret != 0 {
+			return ret
+		}
+	}
+	return 0
+}
+
+// BuildRangesByRow will build range of the given row. It will eval each function's arg then call BuildRange.
+func (cwc *ColWithCmpFuncManager) BuildRangesByRow(ctx *rangerctx.RangerContext, row chunk.Row) ([]*ranger.Range, error) {
+	exprs := make([]expression.Expression, len(cwc.OpType))
+	exprCtx := ctx.ExprCtx
+	for i, opType := range cwc.OpType {
+		constantArg, err := cwc.opArg[i].Eval(exprCtx.GetEvalCtx(), row)
+		if err != nil {
+			return nil, err
+		}
+		cwc.TmpConstant[i].Value = constantArg
+		newExpr, err := expression.NewFunction(exprCtx, opType, types.NewFieldType(mysql.TypeTiny), cwc.TargetCol, cwc.TmpConstant[i])
+		if err != nil {
+			return nil, err
+		}
+		exprs = append(exprs, newExpr) // nozero
+	}
+	// We already limit range mem usage when buildTemplateRange for inner table of IndexJoin in optimizer phase, so we
+	// don't need and shouldn't limit range mem usage when we refill inner ranges during the execution phase.
+	ranges, _, _, err := ranger.BuildColumnRange(exprs, ctx, cwc.TargetCol.RetType, cwc.ColLength, 0)
+	if err != nil {
+		return nil, err
+	}
+	return ranges, nil
+}
+
+func (cwc *ColWithCmpFuncManager) ResolveIndices(schema *expression.Schema) (err error) {
+	for i := range cwc.opArg {
+		cwc.opArg[i], err = cwc.opArg[i].ResolveIndices(schema)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// String implements Stringer interface.
+func (cwc *ColWithCmpFuncManager) String() string {
+	buffer := bytes.NewBufferString("")
+	for i := range cwc.OpType {
+		fmt.Fprintf(buffer, "%v(%v, %v)", cwc.OpType[i], cwc.TargetCol, cwc.opArg[i])
+		if i < len(cwc.OpType)-1 {
+			buffer.WriteString(" ")
+		}
+	}
+	return buffer.String()
+}
+
+const emptyColWithCmpFuncManagerSize = int64(unsafe.Sizeof(ColWithCmpFuncManager{}))
+
+// MemoryUsage return the memory usage of ColWithCmpFuncManager
+func (cwc *ColWithCmpFuncManager) MemoryUsage() (sum int64) {
+	if cwc == nil {
+		return
+	}
+
+	sum = emptyColWithCmpFuncManagerSize + int64(cap(cwc.compareFuncs))*size.SizeOfFunc
+	if cwc.TargetCol != nil {
+		sum += cwc.TargetCol.MemoryUsage()
+	}
+	if cwc.AffectedColSchema != nil {
+		sum += cwc.AffectedColSchema.MemoryUsage()
+	}
+
+	for _, str := range cwc.OpType {
+		sum += int64(len(str))
+	}
+	for _, expr := range cwc.opArg {
+		sum += expr.MemoryUsage()
+	}
+	for _, cst := range cwc.TmpConstant {
+		sum += cst.MemoryUsage()
+	}
+	return
+}
