+++ pkg/planner/core/task.go
@@ -32,6 +32,7 @@ import (
 	"github.com/pingcap/tidb/pkg/planner/core/operator/baseimpl"
 	"github.com/pingcap/tidb/pkg/planner/core/operator/logicalop"
 	"github.com/pingcap/tidb/pkg/planner/core/operator/physicalop"
+	"github.com/pingcap/tidb/pkg/planner/core/operator/physicalop/task"
 	"github.com/pingcap/tidb/pkg/planner/funcdep"
 	"github.com/pingcap/tidb/pkg/planner/property"
 	"github.com/pingcap/tidb/pkg/planner/util"
@@ -74,7 +75,7 @@ func attachPlan2Task(p base.PhysicalPlan, t base.Task) base.Task {
 			p.SetChildren(v.indexPlan)
 			v.indexPlan = p
 		}
-	case *RootTask:
+	case *task.RootTask:
 		p.SetChildren(v.GetPlan())
 		v.SetPlan(p)
 	case *MppTask:
@@ -137,7 +138,7 @@ func attach2Task4PhysicalUnionScan(pp base.PhysicalPlan, tasks ...base.Task) bas
 			// set child will substitute original child slices, not an in-place change.
 			p.SetChildren(&clonedSel)
 			p.SetStats(task.Plan().StatsInfo())
-			rt := task.(*RootTask)
+			rt := task.(*task.RootTask)
 			rt.SetPlan(p) // root task plan current is p headed.
 			// shallow clone proj.
 			clonedProj := *pj
@@ -700,7 +701,7 @@ func calcPagingCost(ctx base.PlanContext, indexPlan base.PhysicalPlan, expectCnt
 	return math.Max(pagingCst-sessVars.GetSeekFactor(nil), 0)
 }
 
-func (t *CopTask) handleRootTaskConds(ctx base.PlanContext, newTask *RootTask) {
+func (t *CopTask) handleRootTaskConds(ctx base.PlanContext, newTask *task.RootTask) {
 	if len(t.rootTaskConds) > 0 {
 		selectivity, _, err := cardinality.Selectivity(ctx, t.tblColHists, t.rootTaskConds, nil)
 		if err != nil {
@@ -2306,21 +2307,4 @@ func (t *MppTask) enforceExchangerImpl(prop *property.PhysicalProperty) *MppTask
 	}
 	nt.warnings.CopyFrom(&t.warnings)
 	return nt
-}
-
-// IndexJoinInfo is generated by index join's inner ds, which will build their own index choice based
-// the indexJoinProp pushed down by index join. While index join still need some feedback by this kind
-// choice info to make index join runtime compatible, like IdxColLens will help truncate the index key
-// to construct suitable lookup contents. KeyOff2IdxOff will help ds to quickly locate the index column
-// from lookup contents. Ranges will be used to rebuild the underlying index range if there is any parameter
-// affecting the index join's inner range. CompareFilters will be used to quickly evaluate the last-col's
-// non-eq range.
-// This kind of IndexJoinInfo will be wrapped as a part of CopTask or RootTask, which will be passed upward
-// to targeted indexJoin to complete the physicalIndexJoin's detail: ref:
-type IndexJoinInfo struct {
-	// The following fields are used to keep index join aware of inner plan's index/pk choice.
-	IdxColLens     []int
-	KeyOff2IdxOff  []int
-	Ranges         ranger.MutableRanges
-	CompareFilters *physicalop.ColWithCmpFuncManager
 }
