+++ pkg/planner/core/task_base.go
@@ -33,287 +33,9 @@ import (
 )
 
 var (
-	_ base.Task = &RootTask{}
 	_ base.Task = &MppTask{}
 	_ base.Task = &CopTask{}
 )
-
-type simpleWarnings struct {
-	warnings []*context.SQLWarn
-}
-
-// WarningCount returns the number of warnings.
-func (s *simpleWarnings) WarningCount() int {
-	return len(s.warnings)
-}
-
-// Copy implemented the simple warnings copy to avoid use the same warnings slice for different task instance.
-func (s *simpleWarnings) Copy(src *simpleWarnings) {
-	warnings := make([]*context.SQLWarn, 0, len(src.warnings))
-	warnings = append(warnings, src.warnings...)
-	s.warnings = warnings
-}
-
-// CopyFrom copy the warnings from src to s.
-func (s *simpleWarnings) CopyFrom(src ...*simpleWarnings) {
-	if src == nil {
-		return
-	}
-	length := 0
-	for _, one := range src {
-		if one == nil {
-			continue
-		}
-		length += one.WarningCount()
-	}
-	s.warnings = make([]*context.SQLWarn, 0, length)
-	for _, one := range src {
-		if one == nil {
-			continue
-		}
-		s.warnings = append(s.warnings, one.warnings...)
-	}
-}
-
-// AppendWarning appends a warning to the warnings slice.
-func (s *simpleWarnings) AppendWarning(warn error) {
-	if len(s.warnings) < math.MaxUint16 {
-		s.warnings = append(s.warnings, &context.SQLWarn{Level: context.WarnLevelWarning, Err: warn})
-	}
-}
-
-// AppendNote appends a note to the warnings slice.
-func (s *simpleWarnings) AppendNote(note error) {
-	if len(s.warnings) < math.MaxUint16 {
-		s.warnings = append(s.warnings, &context.SQLWarn{Level: context.WarnLevelNote, Err: note})
-	}
-}
-
-// GetWarnings returns the internal all stored warnings.
-func (s *simpleWarnings) GetWarnings() []context.SQLWarn {
-	// we just reuse and reorganize pointer of warning elem across different level's
-	// task warnings slice to avoid copy them totally leading mem cost.
-	// when best task is finished and final warnings is determined, we should convert
-	// pointer to struct to append it to session context.
-	warnings := make([]context.SQLWarn, 0, len(s.warnings))
-	for _, w := range s.warnings {
-		warnings = append(warnings, *w)
-	}
-	return warnings
-}
-
-// ************************************* RootTask Start ******************************************
-
-// RootTask is the final sink node of a plan graph. It should be a single goroutine on tidb.
-type RootTask struct {
-	p base.PhysicalPlan
-
-	// For copTask and rootTask, when we compose physical tree bottom-up, index join need some special info
-	// fetched from underlying ds which built index range or table range based on these runtime constant.
-	IndexJoinInfo *IndexJoinInfo
-
-	// warnings passed through different task copy attached with more upper operator specific warnings. (not concurrent safe)
-	warnings simpleWarnings
-}
-
-// GetPlan returns the root task's plan.
-func (t *RootTask) GetPlan() base.PhysicalPlan {
-	return t.p
-}
-
-// SetPlan sets the root task' plan.
-func (t *RootTask) SetPlan(p base.PhysicalPlan) {
-	t.p = p
-}
-
-// Copy implements Task interface.
-func (t *RootTask) Copy() base.Task {
-	nt := &RootTask{
-		p: t.p,
-
-		// when copying, just copy it out.
-		IndexJoinInfo: t.IndexJoinInfo,
-	}
-	// since *t will reuse the same warnings slice, we need to copy it out.
-	// because different task instance should have different warning slice.
-	nt.warnings.Copy(&t.warnings)
-	return nt
-}
-
-// ConvertToRootTask implements Task interface.
-func (t *RootTask) ConvertToRootTask(_ base.PlanContext) base.Task {
-	// root -> root, only copy another one instance.
-	// *p: a new pointer to pointer current task's physic plan
-	// warnings: a new slice to store current task-bound(p-bound) warnings.
-	// *indexInfo: a new pointer to inherit the index join info upward if necessary.
-	return t.Copy().(*RootTask)
-}
-
-// Invalid implements Task interface.
-func (t *RootTask) Invalid() bool {
-	return t.p == nil
-}
-
-// Count implements Task interface.
-func (t *RootTask) Count() float64 {
-	return t.p.StatsInfo().RowCount
-}
-
-// Plan implements Task interface.
-func (t *RootTask) Plan() base.PhysicalPlan {
-	return t.p
-}
-
-// MemoryUsage return the memory usage of rootTask
-func (t *RootTask) MemoryUsage() (sum int64) {
-	if t == nil {
-		return
-	}
-	sum = size.SizeOfInterface + size.SizeOfBool
-	if t.p != nil {
-		sum += t.p.MemoryUsage()
-	}
-	return sum
-}
-
-// AppendWarning appends a warning
-func (t *RootTask) AppendWarning(err error) {
-	t.warnings.AppendWarning(err)
-}
-
-// ************************************* RootTask End ******************************************
-
-// ************************************* MPPTask Start ******************************************
-
-// MppTask can not :
-// 1. keep order
-// 2. support double read
-// 3. consider virtual columns.
-// 4. TODO: partition prune after close
-type MppTask struct {
-	p base.PhysicalPlan
-
-	partTp   property.MPPPartitionType
-	hashCols []*property.MPPPartitionColumn
-
-	// rootTaskConds record filters of TableScan that cannot be pushed down to TiFlash.
-
-	// For logical plan like: HashAgg -> Selection -> TableScan, if filters in Selection cannot be pushed to TiFlash.
-	// Planner will generate physical plan like: PhysicalHashAgg -> PhysicalSelection -> TableReader -> PhysicalTableScan(cop tiflash)
-	// Because planner will make mppTask invalid directly then use copTask directly.
-
-	// But in DisaggregatedTiFlash mode, cop and batchCop protocol is disabled, so we have to consider this situation for mppTask.
-	// When generating PhysicalTableScan, if prop.TaskTp is RootTaskType, mppTask will be converted to rootTask,
-	// and filters in rootTaskConds will be added in a Selection which will be executed in TiDB.
-	// So physical plan be like: PhysicalHashAgg -> PhysicalSelection -> TableReader -> ExchangeSender -> PhysicalTableScan(mpp tiflash)
-	rootTaskConds []expression.Expression
-	tblColHists   *statistics.HistColl
-
-	// warnings passed through different task copy attached with more upper operator specific warnings. (not concurrent safe)
-	warnings simpleWarnings
-}
-
-// Count implements Task interface.
-func (t *MppTask) Count() float64 {
-	return t.p.StatsInfo().RowCount
-}
-
-// Copy implements Task interface.
-func (t *MppTask) Copy() base.Task {
-	nt := *t
-	// since *t will reuse the same warnings slice, we need to copy it out.
-	// cause different task instance should have different warning slice.
-	nt.warnings.Copy(&t.warnings)
-	return &nt
-}
-
-// Plan implements Task interface.
-func (t *MppTask) Plan() base.PhysicalPlan {
-	return t.p
-}
-
-// Invalid implements Task interface.
-func (t *MppTask) Invalid() bool {
-	return t.p == nil
-}
-
-// ConvertToRootTask implements Task interface.
-func (t *MppTask) ConvertToRootTask(ctx base.PlanContext) base.Task {
-	return t.Copy().(*MppTask).ConvertToRootTaskImpl(ctx)
-}
-
-// MemoryUsage return the memory usage of mppTask
-func (t *MppTask) MemoryUsage() (sum int64) {
-	if t == nil {
-		return
-	}
-
-	sum = size.SizeOfInterface + size.SizeOfInt + size.SizeOfSlice + int64(cap(t.hashCols))*size.SizeOfPointer
-	if t.p != nil {
-		sum += t.p.MemoryUsage()
-	}
-	return
-}
-
-// AppendWarning appends a warning
-func (t *MppTask) AppendWarning(err error) {
-	t.warnings.AppendWarning(err)
-}
-
-// ConvertToRootTaskImpl implements Task interface.
-func (t *MppTask) ConvertToRootTaskImpl(ctx base.PlanContext) (rt *RootTask) {
-	defer func() {
-		// mppTask should inherit the indexJoinInfo upward.
-		// because mpp task bottom doesn't form the indexJoin related cop task.
-		if t.warnings.WarningCount() > 0 {
-			rt.warnings.CopyFrom(&t.warnings)
-		}
-	}()
-	// In disaggregated-tiflash mode, need to consider generated column.
-	tryExpandVirtualColumn(t.p)
-	sender := PhysicalExchangeSender{
-		ExchangeType: tipb.ExchangeType_PassThrough,
-	}.Init(ctx, t.p.StatsInfo())
-	sender.SetChildren(t.p)
-
-	p := PhysicalTableReader{
-		tablePlan: sender,
-		StoreType: kv.TiFlash,
-	}.Init(ctx, t.p.QueryBlockOffset())
-	p.SetStats(t.p.StatsInfo())
-	collectPartitionInfosFromMPPPlan(p, t.p)
-	rt = &RootTask{}
-	rt.SetPlan(p)
-
-	if len(t.rootTaskConds) > 0 {
-		// Some Filter cannot be pushed down to TiFlash, need to add Selection in rootTask,
-		// so this Selection will be executed in TiDB.
-		_, isTableScan := t.p.(*physicalop.PhysicalTableScan)
-		_, isSelection := t.p.(*physicalop.PhysicalSelection)
-		if isSelection {
-			_, isTableScan = t.p.Children()[0].(*physicalop.PhysicalTableScan)
-		}
-		if !isTableScan {
-			// Need to make sure oriTaskPlan is TableScan, because rootTaskConds is part of TableScan.FilterCondition.
-			// It's only for TableScan. This is ensured by converting mppTask to rootTask just after TableScan is built,
-			// so no other operators are added into this mppTask.
-			logutil.BgLogger().Error("expect Selection or TableScan for mppTask.p", zap.String("mppTask.p", t.p.TP()))
-			return base.InvalidTask.(*RootTask)
-		}
-		selectivity, _, err := cardinality.Selectivity(ctx, t.tblColHists, t.rootTaskConds, nil)
-		if err != nil {
-			logutil.BgLogger().Debug("calculate selectivity failed, use selection factor", zap.Error(err))
-			selectivity = cost.SelectionFactor
-		}
-		sel := physicalop.PhysicalSelection{Conditions: t.rootTaskConds}.Init(ctx, rt.GetPlan().StatsInfo().Scale(selectivity), rt.GetPlan().QueryBlockOffset())
-		sel.FromDataSource = true
-		sel.SetChildren(rt.GetPlan())
-		rt.SetPlan(sel)
-	}
-	return rt
-}
-
-// ************************************* MPPTask End ******************************************
 
 // ************************************* CopTask Start ******************************************
